feat: MultiOutlet & Ticket storage WIP
- Migrate multi ticketStorage & multiOutlet to replace existing implementation
- Various fixes
- Comment out code where the implementation is not completed
